新增资源使用目录代码。
改动：
1.新资源使用目录代码而不是全路径。
2.错误码规范改动。

已知bug:
1.资源服务不能正确处理路径中有win盘符的情况。

+++ dataaccess/src/main/java/com/sheepfly/media/dataaccess/entity/Resource.java
@@ -94,5 +94,5 @@ public class Resource implements Serializable, EntityInterface, LogicDelete {
      * <p>{@link Directory}</p>
      */
     @Column(name = "DIR_CODE", nullable = false)
-    private Integer dirCode;
+    private Long dirCode;
 }
